display: fb: use a single plane to describe the packed RGB frame buffer
No need to allocate a picture for that. It was not allocating any data, we
forced the one plane values anyway.

We can use plane_CopyPixels() directly on that single plane.

When we know the bits_per_pixel we request the core the matching packed RGB
format with the same width/depth.

If the forced chroma is not a packed/single plane chroma the plane copy won't
be sufficient, but that's a pre-existing issue. Only one plane is currently
supported.