vlc_es: add ICC profile struct
We store the ICC profile as part of the (ancillary) data allocation,
using a trailing unsized array. This is because we need to copy it in
and out of various libraries anyway, to avoid having to expose things
like AVBufferRef.

The (as of writing) current ICC specification can be found at
https://www.color.org/specification/ICC.1-2022-05.pdf