jpeg: decode ICC profiles
This implementation relies on the helper function jpeg_read_icc_profile
which was introduced in libjpeg-turbo v1.5.90. Its absence from other
versions of libjpeg means we may wish to copy this helper function into
our own code in the future. But for now, keep it simple and avoid
unnecessary bloat.