xcb/window: wait for window to be mapped
The user of the window typically uses a different connection to the X11
than the window module. The window must be mapped before the user tries
anything with it. On a different connection a mere `xcb_flush()` is not
sufficient, as it does not guarantee that the X11 server served the
request.

The current code relies on the fact that the window map request will
necessarily have been processed before the pointer query reply is
received. But the pointer query will be (re)moved, so it cannot be
relied on any longer.