video_output: remove dropped_current_frame test
When we go in the `(next != NULL)` section, we have a new picture to render.
We need to wait until its "refresh" date to start the rendering.
It doesn't matter if there was a picture shown before that one.

The code is almost the same as before where dropped_current_frame was almost
always true. The only time it was false was:
* first picture and next is not NULL (first pic & next NULL is impossible here)
* not first picture and next is NULL

So the new code only changes the case where it is the "first picture and next
is not NULL". But in that case, first was true and therefore the test was
not passing, just like now.

It is logically equivalent to the code before. But now it's clear that when we
don't have a next picture we request a redisplay (of displayed.current).