wall: fix non-first abscissa of non-8-bit samples
The X and Y values are pixel coordinates in the potentially subsampled
plane resolution. This must be converted to an octet offset as the type
of the pixel plane pointer is pointer-to-uint8_t.

So the X value must be multiplied by the sample size. The current code
was instead rounding the value down to a multiple of the sample size.

Bug reported by Philipp Marek on IRC.

(cherry picked from commit 87733c5011912e3f46e48948b0fe116b72bfa83f)