vout : apple : Prevent a crash on draw appearing on tvOS
It appeared the EAGL context buffer allocation fails on tvOS when the display isn't connected.
Given the attached render buffer was incomplete, it led to a crash when a draw was attempted by the renderer.
There are two locations where this patch attempts to improve the situation by prepraring the buffers if a failure was previously encountered :

- As the UIApplicationDidBecomeActiveNotification is dispatched when a screen connection is happening on tvOS, we observe it to restore the render buffers if needed.
- We also try to restore the render buffers when the drawing context is requested from make_current