xcb/window: track and report ICC profile changes
We need to update the ICC profile either if the window moves to a new
display, or if the root window atom associated with a given display is
updated. To this end, we need to start tracking property changes on the
root window, as well as updating this atom every time the window moves
to a new display.

Unfortunately, this code again requires one memcpy more than I'd like to
have there, as a result of xcb not letting me fetch properties into my
own buffers. If we determine this to be a performance issue later on,
the straightforward fix would be to make vlc_icc_profile_t refcounted
with a custom free callback.