input: decoder: extract substream handling
The substream handling was under a lot of precondition, leading to a lot
of indentation. Moving this handling into a separate function allows
early return within the function, which reduce to a single indentation
level and greatly simplify the reading of the function while simplifying
the caller site where it was only a specific case to handle.