transcode: add the pcr_sync utility
This utility helps handling PCR events in modules where frames are
treated asynchronously. This tool basically bufferize PCR events until
all associated frames are out of the async unit.

The helper was made to handle any async frame treatment unit and could
be used in any other modules that have the same behavior. For now, only
transcode is in that case so the files will reside in the transcode
folder.