messages: simplify Win32DebugOutputMsg()
No need to call vsnprintf() twice and add the ending \n in an extra part of the
temporary buffer. We have all that we need with a single vasprintf() call.

No need to make a va_copy() inside Win32DebugOutputMsg(), there is one on the
caller side and now we only use the va_list once.