test: add iosvlc.m for iOS development
iosvlc.m provides a binary usable as an iOS application, forwarding the
VLC arguments just like VLC on desktop. It allows easier iteration on
vlccore development for iOS, without the need to test in a VLCKit
application like VLC for iOS or new external application.

It is currently designed for usage with dynamic plugins.

To develop with it, you must generate a .ipa archive containing both
the resulting binary as executable, a PkgInfo file, an Info.plist file
describing the package and the libs (libvlc.dylib, libvlccore.dylib, and
every plugin .dylib or additional convenience libraries that are not
linked statically in the Frameworks/ directory. It must then be signed
with a developer certificate allowed by Apple and provisionned with a
mobileprovision file allowing installation on the given device for the
same developer certificate.

Then, tools like libimobiledevice can be used to start the application
with additional arguments or environment variables. They can also be
added in XCode through the "Edit Scheme" menu.

A big part of the iOS-specific code has been originally written by
Marvin Scholz in a more complete libVLC ios sample.

Co-authored-by: Marvin Scholz <epirat07@gmail.com>

Cherry-picked and adapted from commits:
e16f4d9bb6830fd5ad257a60babc0377a3ab39a2
0bfe88030104d2e9c962c24c144ce5c17294f146
96bee41831849597448a2414aa06faa1ee837a38
904998afec0b2d4192e877c19c7e5295bcaf4c67
54c392d7be7f73af6fabcf542c3c2670db00ff1a