input: decoder: move reset of i_preroll_end
The i_preroll_end state is reset at the end of the flush call on the
es_out side, but it's only set from the decoder thread which would be
flushing the decoders anyway, so it can be set from there (ie. the last
point of flush) where it's already locked.