contrib: update vulkan-loader to 1.3.211
This is needed as a consequence of the update to vulkan-headers, as the
older loader does not compile against too-new headers. Pick the latest
version supported by the MSYS2 patches at the time of commit, and also
update said patches.

Somehow the static library built changed from libVKstatic.a to
libvulkan.dll.a. I have no idea how this mishmash of extensions comes
about, but I've simply adapted the name to get it to compile.