contrib: drop libplacebo vulkan linux exception
This logic is non-functional on recent libplacebo, which always builds
vulkan support as long as vulkan headers are available (which it is on
raspbian).

More importantly, excluding linux also breaks android builds, because
android's built-in copy of the vulkan headers (erroneously) omits
vk.xml. So we need to build vulkan-headers ourselves on this platform.