libplacebo: switch to new v4 overlay API
This is not exactly too exciting for us, because we currently only have
overlays consisting of a single region per picture_t. But in theory,
this new API allows us to have multiple regions (parts) pointing at the
same subpicture.

It's also a mandatory change, as the old API will be dropped in
libplacebo v5.