HACK: modules: bank: load framework plugins on iOS
On iOS, there's no folder plugins, and every plugins are located at the
Application root/Frameworks/ folder, so we need to probe this folder
instead of the plugins/ suffixed path. In addition, the library path
located by VLC will redirect to foo.app/Frameworks/vlccore.framework/vlc
which doesn't correspond to the correct path. It mostly affects 3.0.x
branch and not next 4.0 or more branches.