win32: thread: apply the vlc_testcancel() code in local calls
And then simplify the code locally.

Now there's only one current_thread_ctx TLS lookup for the whole duration of
vlc_tick_wait().

The killed variable is only check if it's a VLC thread and killable is true.

Not sure why the Win8+ code differs from the rest, even the memory order
loading of th->killed is different.