vout: correctly track ICC updates in vout_display_t
The line that updated the current `vd->cfg` got accidentally removed in
a previous refactor of this commit, leading to this value being
out-of-date. This broke any vouts which directly looked at
`vd->cfg->icc_profile` instead of taking the current profile from the
callback itself.

Also add a clarifying comment to this function, since it just blindly
copies over a pointer. In practice, the `vout_thread_t` takes care of
this.