win32: thread: remove "wall" clock source option
It's slower than the "perf" clock source and doesn't provide any extra
precision. We don't need the UTC reference it provides.

We keep it as the default source until the user/default preferences are applied
since it provides an good granularity in a time where things may happen fast
(libvlc startup). The 10/16 millisecond granularity of the other sources are
not very useful there.