win32: thread: bring back the winmm clock source
It is the default on 3.0 and might be useful when QueryPerformanceCounter() is
not accurate or too slow to call.

* https://docs.microsoft.com/en-US/troubleshoot/windows-server/performance/programs-queryperformancecounter-function-perform-poorly
* https://www.virtualdub.org/blog2/entry_106.html
* https://chromium.googlesource.com/chromium/src/base/+/refs/heads/main/time/time_win.cc

Partially reverts fdbbae70af8005bcaa771ef63968a2d023da267e. Now we link
statically with winmm. It should be on all Win7+ machines.

We also don't force a period value smaller than wPeriodMin.