contrib: qt: enforce the creation of a .pc file
Ensure the files are generated with a pkgconfig file. In particular, we
want Qt plugins and QML plugins to have a .pc file so as to find them
and link them from configure.ac.

Note that the generated plugin file will not be valid currently, and in
particular Libs: will wrongly point to $(PREFIX)/lib although the
library is installed in $(PREFIX)/plugins/$$[PLUGIN_TYPE].

Also enforce VERSION for the plugins. Without enforcing VERSION, the
QMAKE_PKGCONFIG_VERSION is not set at all on Windows, and Version: in
the .pc file is not present. It leads to invalid pkgconfig files on
Windows.