youtube.lua: fix crash on unexpected "n" descrambling transformation code
It was possible, when encountering different code from what was expected
for some known transformations, to void the code parsing pointer instead
of advancing it, resulting in a subsequent crash of the script and total
playback failure. Add a fallback and check, to prevent and gracefully
deal with this, and still allow playback, even if throttled, in case of
descrambling failure.

(cherry picked from commit 149444107513953b46f52675cea0408f324d4c04)
Signed-off-by: David Fuhrmann <dfuhrmann@videolan.org>