qt: register plugins for wayland
They are currently enabled only when the matching contrib has been found
on the system. The plugin name doesn't match with the plugin from the
contrib, but still enforce at link time the plugin to be present.

This always use the XdgShell integration regardless of what is supported
for now.