decklink: fix fourcc string to enum conversion
Before 9168091a46266f51e30e49045627b9d23ff75e0e htonl() was used to convert
the values to/from enum to the string. But that commit broke the conversion on
little endian machines.

Now we don't use an intermediate string anymore. We just use GetDWBE() to do
the endianness conversion of the char buffer if needed.