contrib: libplacebo: force the C++ runtime linking not found in vulkan-loader
When we're not building vulkan-loader the .pc file may not be found at all or
it may contain the C++ runtime export. To properly use libplacebo from C code
we need to link with the C++ runtime of vulkan-loader.

It's not a good idea to rebuild vulkan-loader on behalf of the system which
contains it with its own settings. So we set the C++ runtime in the library
that uses it.