contrib: asdcplib: Force building with a specific older C++ version
This project uses the 'register' storage class specifier for a lot
of variables, and C++17 disallows that storage class spcifier.

This fixes building with newer compilers that default to C++17
if nothing is specified (which is the default since GCC 11
and Clang 16).

When building with Clang 16 in C++17 mode, it fails with errors
like these:

KM_util.cpp:357:3: error: ISO C++17 does not allow 'register' storage class specifier [-Wregister]
  register byte_t c = 0, d = 0;
  ^~~~~~~~~