mmdevice: add a timer API
Use "aout" internal threading to implement a timer.

This avoid to create an extra thread/timer.

This will also simplify locking since aout_stream callbacks,
aout_stream_timer_Trigger() and the timer callback need to be protected
by a mutex.

Furthermore, this leaves the MTA handling to the aout module.