contrib: qt: fix .pc generation for qml_plugins
The qml_plugins did not include a rewrite rule like those that were
written in 09e3cc77daeaf597a53f5a3937c386ea1e73e47c. Consequently, the
.pc files were invalid and linking to the contrib build folder instead
of the installation prefix folder. This commit is the same as the
previous commit but dedicated to qml_plugins specifically, instead of
just qt_plugins.