contrib: don't rebuild meson projects if nothing changed
Removing only the meson-private data will trigger a reconfiguration of the
project. If the configuration ends up being the same as before only the
modified files will be rebuilt.

This is similar to the configure we redo with autotools targets. We don't
rebuild everything if the configure produced the same configuration. If source
files have changed they will be rebuilt.