contrib: use the variables similar to emcmake for emscripten
emcmake is supposed to be used with CMake and emscripten [1] and it uses a
toolchain file for that [2]. But we already have our own with more things. So let's
use the same toolchain file in [3] which may overwrite some of our own values with
the ones that are expected for CMake emscripten.

This should allow CheckTypeSize and TestBigEndian tests to work properly.

[1] https://badlydrawnrod.github.io/posts/2020/05/19/emcmake-with-emscripten/
[2] https://github.com/emscripten-core/emscripten/blob/main/emcmake.py
[3] https://github.com/emscripten-core/emscripten/blob/main/cmake/Modules/Platform/Emscripten.cmake