test: iosvlc: provide asset:// redirection
asset:// scheme will be redirected to file:// access with the correct
bundle path setup as prefix, so that any asset:// resource is accessed
from the application package directly.

With EXTRA_FILE=sample.mp4 ../extras/package/apple/bundle.sh, it
provides the ability to play the file by running the application with
the argument `asset://sample.mp4`.