test: input_decoder: send preliminary pictures
Send initial pictures to start-up the whole input_decoder state and
ensure it won't deadlock the es_out.

At the same time, it seems that updating the vout state from the decoder
(decoder_UpdateVideoOutput) is racy with the work of the video output
itself, so ensure the update is done once at the beginning for now. A
subsequent test will be provided once this race is fixed.