video_output: check picture != NULL before release
In the UpdateCurrentPicture function, it looks like the picture will
never be NULL because we check at the beginning of the function whether
it's NULL or not and only set it at the end.

However, UpdateCurrentPicture calls PreparePicture to make progress, and
this last function can call ChangeFilters, leading to the call to
FilterFlush. After this sequence, FilterFlush will set the picture
syys->displayed.current to NULL and the code will unwind back to the
UpdateCurrentPicture call, ready to release the picture.

PreparePicture should probably be made so that it doesn't reset the
current state, but it's far simpler to first fix the problem with a
check after PreparePicture, guarding picture_Release from NULL pointers
and then test the behaviour so that refactor doesn't break it.