qt: refactor media tree listener
This C++ listener is actually a wrapper over a media tree listener, not
a media source listener.

This allows to make NetworkMediaModel independent of media sources, so
that it can be reused with only a media tree not linked to a media
source.

NetworkDeviceModel, which still requires a media source, stores it in
its listener callbacks implementation (ListenerCb).