transcode: pcr_sync: change the PCR fast-forwarding
Fast-forwarding a PCR means that the pcr_sync immediately tells the API
user that he can send it in the case there is no frame currently being
watched.
This happen in three scenarios:
  - The first SetPCR call is always fast-forwarded because no data
    should have entered the pcr_sync yet.
  - The stream has a "data-hole" and no frames enters it (most usual
    case being subtitles) while PCR keeps coming. This ensure PCR values
    are still forwarded even when the frame queue is empty.
  - The encoder has no delay. Ie: A frame is immediately encoded and
    returned. All PCR values must then be forwarded.

The last point here wasn't handled correctly due to bad implementation
and was a blocker for no-delay encoder (subtitles or some audio ones).

Now, we check if the frame input and output are on the same page before
queuing a PCR event and if it's the case, the PCR value is simply
fast-forwarded.

A following commit will add test cases for this usage.