contrib: speexdsp: disable NEON just once
On AArch64, NEON is unconditionally disabled right above. This was only
adding it a second time on all AArch64 targets other than iOS and
Android, with no real effects.

On non-ARM architectures, the flag is useless, and HAVE_NEON is
meaningless so ignore it.

On AArch32, HAVE_NEON is still abode by; this patch is a no-op.