contrib: speexdsp: fix NEON support on Linux
The HAVE_NEON flag is only ever defined on select platforms. That is to
say first Android which had one build with and one without it. Later on,
it was also added to iOS.

It cannot be assumed that NEON is not there just because the flag is
unspecified. This silently broke NEON support on Linux, and literally
any other platform at the time this regression was introduced.

Strictly speaking, we could keep checking the flag on Android or iOS on
ARM or AArch64. But as of now, all four of those platforms always set
HAVE_NEON, so this would not have any effect. We still forcefully
disable NEON on Android-x86, which was the original purpose of this
stanza.

Regression from ef98e81ff014175fd7494c8313de8883d15834b3.