contrib: assume iOS armv7 targets have NEON support
This is strictly identical to the previous situation since HAVE_NEON is
set during bootstrapping on iOS only for armv7 targets anyway.

The commit prepares the removal of the HAVE_NEON flag completely in later
commits, which is redundant with armv7 check since we don't support
non-neon-capable armv7 targets anymore.