extras: tools: Tidy up cmake-enable-ALPN-support-on-macOS-10.14.patch
In b8eb8c6d4ea7dc8f8707550fa6c05b3e3ccb62fc, this patch was manually
touched up to make it apply on a newer version of CMake - the line
counts mismatched slightly, but GNU patch 2.7.6 still happily
managed to apply it. However patch 2.5.8 in macOS failed to apply
the patch due to the mismatched numbers of lines.

This makes the patch apply correctly with the version of patch
on macOS too.

This should fix issue #27381.