ci: Disable branch protection for the current Windows/ARM64 builds
Clang does support the flag and gladly generates the instructions
for return address signing, but it doesn't generate matching unwind
info for it.

If a return address is signed, unwinding through it would fail unless
the signing is reversed at the right place. The ARM64 SEH unwind info
format has no public documentation on how to handle return address
signing currently.

This issue isn't noticeable on any current public Windows on ARM64
devices or servers, as the CPUs don't implement the return address
signing instructions (which then are handled as no-ops). However,
binaries with those instructions would fail subtly (seemingly work
fine, but fail whenever unwinding the stack) on newer CPUs.

Therefore, disable it in the CI configuration, as long as that uses
a version of Clang that doesn't generate correct unwind info for this
case.