wasm-emscripten: add support for address sanitizer
In this commit, the --with-sanitizer option is forwarded to the
configure script for VLC, but also intercepted so that the emcc call
creating the plugin list can use the option as well.

It's necessary for now since the plugin list is not generated by
automake yet, thus the call to emcc is not intercepted to add the
sanitizer options added from VLC's configure.ac.