contrib: qt: use system zlib
This was attempted in 4f72f199f20530803227ca09478d6004a0bb9703, but
back then, qtsvg still needed the Qt vendored zlib. Now, this has been
fixed and we don't need the vendored zlib. zlib from the contribs has a
few fixes [1] over the Qt version, so it is prefered.

Just like the explicit -qt-zlib, we specify -system-zlib so we don't
rely on the automatic fallback to the vendored version. Since this fails
at configure time, it has been helpful when hacking away at the compiler
flags and options.

1. https://github.com/madler/zlib/blob/master/ChangeLog