contrib: remove stray HAVE_NEON
This was originally added to make builds with and without NEON on
Android. I am not sure why this was at all needed as Android supports
run-time detection using the normal Linux HWCAP mechanism, so a single
build should have been feasible fairly easily.

But anyhow, this is no longer used anywhere.