opengl: make placebo conditional on <libplacebo/opengl.h>
I want to start using the pl_opengl integration, so we need to slightly
revise the check to ensure we also have the libplaceo opengl header
available.

It's worth pointing out that these headers eventually become
non-conditional (in libplacebo git master as of writing), but we need to
live with this transitional logic until then.