opengl: persist pl_shader_obj until shader is freed
This code, as written, is technically undefined behavior. It was fine so
far because we didn't end up using the shader objects in any way that
wasn't baked into the resulting pl_shader_res, but it needs to be fixed
to allow shaders to reference GPU resources which may be used later on.