opengl: change libplacebo var type check to assertion
Simply continuing in this error case without assigning the corresponding
uniform will silently lead to a broken shader / image. Since we rely on
these assumptions, we should assert them so any such bugs are actually
found in a debuggable way.

This is only a theoretical risk since, as the comment points out, we
don't rely on any such variables. But maybe this will unexpectedly
change in the future.