sout: add a tracer utility stream filter
The VLC tracing API is really useful in a lot of debugging scenario. It
was previously implemented in playback but lack an implementation in the
stream output chain.

This patch propose a named stream filter that can be placed before any
other module and trace their input. It allow the user to choose
precisely at which node of the chain they want tracing and under which
name.

An example usage can be found in the help section of the module:

```sout
```