opengl: support .cube LUTs
To keep the API and integration as simple as possible, only support
"conversion" style LUTs, i.e. LUTs which are intended to map from the
source color space to the display color space in one step.

This represents the most common type of LUT in practice, and in
particular, the sample files that have been requested by users.