contrib: don't consider windows-gnu as mingw32
windows-gnu is the name of the msys environment, which is a UNIX like
environment that runs in Windows. mingw32 is an actual Windows environment. The
msys installer provides 3 environment: windows-gnu, mingw32 and mingw64.
Depending which you use will make "gcc" produce different dumpmachine values:

- x86_64-pc-msys
- x86_64-w64-mingw32
- i686-w64-mingw32

Hiding windows-gnu means in an msys environment we do not use the HOST- prefix
and compile for msys rather than for Windows, when using
--host=x86_64-w64-mingw32. If one wants to build VLC for msys, one should use
--host=x86_64-pc-msys and the build system will match.

We already assume MSYS_BUILD in all shell environment where the build string
contains msys or mingw32.