aout: fix drift handling when using aout_TimingReport()
Drift handling was not fully tested when the aout was using
aout_TimingReport() (and the audio was not the master).

It was working for big drifts (silence or flush) but not for small ones,
when the resampler was used because the last set drift was reset once
used.

Furthermore, stream_Synchronize() should not use past points but up to
date points including:
 - The silence already sent
 - The block resize (bigger or smaller) due to the resampler

The simplest way to achieve that is to use the audio delay, that is now
fixed when using aout_TimingReport().