contrib: add python virtual env support
Some packages depend on python packages being available at compile time.
This requires the use of a python virtual environment, to make these
dependencies available to the corresponding commands.

This commit adds support for creating a venv via the new .python-venv
target that can be added as a dependency of bundled python packages.